Using custom fields to preserve ENUM types from postgres schema
I was not aware that the guacamole client (the war file) would use the
custom ENUM types at the top of the file (as opposed to just checking
for string equality).

To preserve these, I utilized custom fields based on this presentation
here:
https://wiki.postgresql.org/images/0/0b/Postgresql_django_extensions.pdf

Each of the models switched out CharFields for these EnumFields.

Because A) we could not automatically migrate the data and B) no one is
using (successfully) this project yet, we included two migrations, one
(0002) to remove the offending fields, and another (0003) to add in the
correct ones.

Because of the method by which we add the actual ENUM types themselves
in 0003, we needed to include sqlparse in the requirements.

Finally, our test was buffed up a little bit to prove we can still do
string comparison at the application level.